+++ src/index.js
@@ -1,3 +1,0 @@
-export {default as LinkedIn} from './LinkedIn';
-export  {default as LinkedInPopUp} from './LinkedInPopUp';
-export {default} from './LinkedIn';