+++ src/index.ts
@@ -1,0 +1,3 @@
+export * from './useLinkedIn';
+export * from './LinkedIn';
+export * from './LinkedInCallback';